printf

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: wonjilee <wonjilee@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/05 13:23:31 by wonjin            #+#    #+#             */
-/*   Updated: 2023/01/16 22:58:50 by wonjilee         ###   ########.fr       */
+/*   Updated: 2023/01/17 21:39:14 by wonjilee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,6 @@
 # define FT_PRINTF_H
 # include <unistd.h>
 # include <stdarg.h>
-# include <stdlib.h>
 
 int	ft_printf(const char *format, ...);
 int	print_char(char c);
@@ -27,7 +26,7 @@ int	print_lhex(unsigned int num);
 int	print_uhex(unsigned int num);
 int	print_pcent(void);
 int	ft_itoa(long long num);
-int	ft_lhex(unsigned int num);
-int	ft_uhex(unsigned int num);
+int	ft_lhex(unsigned long num);
+int	ft_uhex(unsigned long num);
 
 #endif
